Admin add lessons

+++ administrator/httprequest/response/getDeleteNews.php
@@ -5,7 +5,9 @@ $id = $_GET['id'];
 $res['success'] = false;
 $gambar = $conn->query("SELECT gambar FROM tbl_news WHERE id = $id")->fetch_assoc()['gambar'];
 if ($conn->query("DELETE FROM tbl_news WHERE id = $id")) {
-    unlink('../../../assets/images/news/' . $gambar);
+    if ($gambar !== 'cryptocurrency1.jpg') {
+        unlink('../../../assets/images/news/' . $gambar);
+    }
     $res['success'] = true;
 }
